requirements.txt
streamlit
Pillow
rembg